+++ base/BootstrapInterface.php
@@ -56,5 +56,5 @@ interface BootstrapInterface
      * Bootstrap method to be called during application bootstrap stage.
      * @param Application $app the application currently running
      */
-    public function bootstrap(Application $app);
+    public function bootstrap($app);
 }