+++ web/Application.php
@@ -56,12 +56,13 @@ class Application extends \yii\base\Application
     /**
      * @inheritdoc
      */
-    public function preloadComponents()
+    protected function bootstrap()
     {
-        parent::preloadComponents();
         $request = $this->getRequest();
         Yii::setAlias('@webroot', dirname($request->getScriptFile()));
         Yii::setAlias('@web', $request->getBaseUrl());
+
+        parent::bootstrap();
     }
 
     /**